* place.py: fix failure of non existent dest

+++ place.py
@@ -31,20 +31,24 @@ def find_attri(f):
     return [n, d]
 
 
+def link_file(f, dest):
+    try:
+        if os.access(os.path.dirname(dest), os.W_OK):
+            if  os.path.exists(dest) and not os.path.islink(dest):
+                shutil.move(dest, dest + '.backup')
+            else:
+                os.unlink(dest)
+        os.symlink(os.path.abspath(f), dest)
+    except OSError:
+        print('*FAILED*: %s => %s' % (f, dest))
+
+
 def place_file(f):
     [fname, d] = find_attri(f)
     if fname and d:
         dest = os.path.expandvars(d.rstrip()+fname.rstrip())
         print('   %s => %s' % (f, dest))
-        try:
-            if os.access(os.path.dirname(dest), os.W_OK):
-                if not os.path.islink(dest):
-                    shutil.move(dest, dest + '.backup')
-                else:
-                    os.unlink(dest)
-            os.symlink(os.path.abspath(f), dest)
-        except OSError:
-            print('*FAILED*: %s => %s' % (f, dest))
+        link_file(f, dest)
     else:
         print('*FAILED*: %s => ???' % f)
